Initial lump commit
Sets up initial project structure, but structure is subject to change.
Includes a setup file.
themis_ra.py includes functionality to parse a json config file into a dictionary,
which should be enough to close #21
utils.py includes some basic utilities that were ripped form the krc project, and
it also houses the functionality to preprocess images.  This functionality should
eventually be moved to a different file, but it should be fine for now.  Closes #22.